test(tests): add celery worker fixture
